Get a single user
