+++ lib/data.js
@@ -8,6 +8,7 @@
 // Dependencies
 const fs = require('fs');
 const path = require('path');
+const { parseJsonToObject } = require('./helpers')
 
 // Container for the module (to be exported)
 const lib = {}
@@ -50,7 +51,10 @@ lib.read = function(dir, file, cb) {
   const dirPath = `${lib.baseDir}${dir}/${file}.json`
 
   fs.readFile(dirPath, 'utf-8', (err, data) => {
-    cb(err, data)
+    if (!err && data) {
+      const parsedData = parseJsonToObject(data)
+      cb(false, parsedData)
+    } else cb(err, data)
   })
 }
