+++ lib/helpers.js
@@ -27,6 +27,19 @@ const helpers = {
     } catch (e) {
       return {}
     }
+  },
+
+  // Validate user required fields
+  valildateFields: payload => {
+    const firstName = typeof(payload.firstName) === 'string' && payload.firstName.trim().length > 0 ? payload.firstName.trim() : false
+    const lastName = typeof(payload.lastName) === 'string' && payload.lastName.trim().length > 0 ? payload.lastName.trim() : false
+    const phone = typeof(payload.phone) === 'string' && payload.phone.trim().length === 11 ? payload.phone.trim() : false
+    const password = typeof(payload.password) === 'string' && payload.password.trim().length > 0 ? payload.password.trim() : false
+    const tosAgreement = typeof(payload.tosAgreement) === 'boolean' && payload.tosAgreement === true ? true : false
+
+    return {
+      firstName, lastName, phone, password, tosAgreement
+    }
   }
 }
